pyproject.toml
[tool.poetry]
name = "wagtailforms"
version = "1.0.0"
description = "Wagtail forms in a streamfield"
authors = ["Lumunix <lumunix@icloud.com>"]
license = "MIT"
readme = "README.rst"
keywords = ["wagtail", "streamfield", "forms","fox-ie"]

[tool.poetry.dependencies]
python = "^3.10"
wagtail = ">=5.2,<=6.1.3"
Unidecode = ">=0.04.14,<2.0"
wagtail-generic-chooser = ">=0.5.0,<0.7"


[tool.poetry.dev-dependencies]
sphinxcontrib-spelling = ">=2.3.0"
sphinx = ">=1.5.2"
sphinx-autobuild = ">=0.6.0"
karma_sphinx_theme = ">=0.0.6"
mock = "==5.1.0"
pytest-django = "==4.8.0"
pytest = "==8.3.2"
isort = {version = "==5.13.2", extras = ["pyproject"]}
flake8 = "==7.1.0"
flake8-blind-except = "==0.2.1"
flake8-debugger = "4.1.2"
wagtail-modeladmin = "^2.0.0"
django-recaptcha = "^4.0.0"
psycopg2-binary = "^2.9.9"
tox = "^4.16.0"
black = "^24.4.2"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length=99
target-version=["py310"]
exclude = '''
/(
    \.git
  | \.tox
  | build
  | dist
  | migrations
)/
'''

